pyproject.toml
[tool.poetry]
name = "llm-jp-eval-mm"
version = "0.1.0"
description = ""
authors = ["Silviase <koki.maeda@nlp.c.titech.ac.jp>", ]
maintainers = ["Silviase <koki.maeda@nlp.c.titech.ac.jp>", ]
readme = ["README.md", "README_en.md"]
packages = [
    {include = "llm_jp_eval_mm", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.10"
datasets = "^2.20.0"
torch = "^2.3.1"
ipykernel = "^6.29.5"
transformers = "^4.42.4"
pillow = "^10.4.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
isort = "^5.13.2"
mypy = "^1.10.1"
black = "^24.4.2"
pyproject-flake8 = "^7.0.0"
types-requests = "^2.32.0.20240622"
huggingface-hub = "^0.23.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py39", "py310", "py311"]

[tool.isort]
line_length = 120
py_version = 310
profile = "black"
src_paths = ["src"]

[tool.flake8]
# E203 is not PEP8 compliant and black insert space around slice (https://black.readthedocs.io/en/stable/faq.html#why-are-flake8-s-e203-and-w503-violated)
# E501: Line too long. Disable it to allow long lines of comments and print lines which black allows.
# W503: line break before binary operator (https://www.flake8rules.com/rules/W503.html)
ignore = ["E203", "F401", "E501", "W503"]
max-line-length = 120

[tool.mypy]
python_version = "3.10"
